Fix MultiIndex

+++ as-packages/chain/assembly/env.ts
@@ -32,7 +32,7 @@ export declare function db_idx64_remove(iterator: i32): void;
 export declare function db_idx64_next(iterator: i32, primary_ptr: usize): i32;
 export declare function db_idx64_previous(iterator: i32, primary_ptr: usize): i32;
 export declare function db_idx64_find_primary(code: u64, scope: u64, table: u64, secondary_ptr: usize, primary: u64): i32;
-export declare function db_idx64_find_secondary(code: u64, scope: u64, table: u64, secondary: u64, primary_ptr: usize): i32;
+export declare function db_idx64_find_secondary(code: u64, scope: u64, table: u64, secondary: usize, primary_ptr: usize): i32;
 export declare function db_idx64_lowerbound(code: u64, scope: u64, table: u64, secondary_ptr: usize, primary_ptr: usize): i32;
 export declare function db_idx64_upperbound(code: u64, scope: u64, table: u64, secondary_ptr: usize, primary_ptr: usize): i32;
 export declare function db_idx64_end(code: u64, scope: u64, table: u64): i32;
